feat: ui default theme (#100)
* Delete unused files and composables

* Delete unused layout and component files

* version(nuxt): release 0.5.0-beta.16

* version(cli): release 0.3.4

* Add branch parameter to defineDownload function

* version(cli): release 0.3.5

* Refactor downloadTemplate function in core.ts

* Add forceClean option to DefineDownloadOptions and DefineDownloadOptions.path

* version(cli): release 0.3.6

* Add existsSync function to prevent unnecessary file copying and add new file to download

* version(cli): release 0.3.7

* version(nuxt): release 0.5.0-beta.17

* Add forceClean option to app.vue template and add pages/index.vue template

* Remove retry delay in defineDownload function

* Update file paths in UI module

* version(cli): release 0.3.8

* version(nuxt): release 0.5.0-beta.18

* Update core.ts and README.yaml, and delete style.css and index.ts

* version(cli): release 0.3.9

* Update pergel version and add pergel.css asset

* version(nuxt): release 0.5.0-beta.19

* version(cli): release 0.3.10

* Refactor folder creation logic in defineDownload function

* fix: lint issues

* Add themes:prepare script to package.json

* Add examples folder to tsconfig.json

* Add "themes" directory to tsconfig.json

* fix: lint issues

+++ packages/cli/src/core.ts
@@ -1,4 +1,4 @@
-import { copyFileSync, rmSync } from 'node:fs'
+import { copyFileSync, existsSync, mkdirSync, rmSync } from 'node:fs'
 import { join, resolve } from 'node:path'
 import { downloadTemplate } from 'giget'
 import { defu } from 'defu'
@@ -30,7 +30,7 @@ export function defineDownload(options: DefineDownloadOptions) {
     }) as DefineDownloadOptions
 
     if (options.file?.dir) {
-      const { source, dir } = await downloadTemplate(join(githubRepo, `${options.file.dir}#${options.branch}`), {
+      const { dir } = await downloadTemplate(join(githubRepo, `${options.file.dir}#${options.branch}`), {
         dir: options.file.tempOutput,
         cwd,
         force: true,
@@ -38,10 +38,39 @@ export function defineDownload(options: DefineDownloadOptions) {
 
       for (const file of options.file.path) {
         const output = resolve(join(cwd, file.outputFileName))
-        copyFileSync(
-          join(dir, file.fileName),
-          resolve(output),
-        )
+
+        if (!existsSync(output)) {
+          // remove last .file extends and name for create folder
+          if (output.split('/').pop()?.includes('.')) {
+            const folder = output.split('/').slice(0, -1).join('/')
+            mkdirSync(resolve(folder), {
+              recursive: true,
+            })
+          }
+
+          copyFileSync(
+            join(dir, file.fileName),
+            resolve(output),
+          )
+        }
+        else if (file.forceClean) {
+          rmSync(output, {
+            recursive: true,
+            force: true,
+          })
+
+          if (output.split('/').pop()?.includes('.')) {
+            const folder = output.split('/').slice(0, -1).join('/')
+            mkdirSync(resolve(folder), {
+              recursive: true,
+            })
+          }
+
+          copyFileSync(
+            join(dir, file.fileName),
+            resolve(output),
+          )
+        }
 
         logger.success(`Downloaded template file: ${output}`)
       }
@@ -51,8 +80,6 @@ export function defineDownload(options: DefineDownloadOptions) {
         force: true,
         retryDelay: 100,
       })
-
-      return { source, dir }
     }
 
     if (options.folder && options.folder.length) {
@@ -61,6 +88,7 @@ export function defineDownload(options: DefineDownloadOptions) {
           dir: folder.output,
           cwd,
           force: true,
+          forceClean: folder.forceClean !== false,
         })
 
         logger.success(`Downloaded template folder: ${dir}`)
